ادرس ها اصلاح شده. هر دفعه درون کگل نیاز به تغییر نیست

+++ lib/models/monodgp/ops/modules/ms_deform_attn.py
@@ -31,7 +31,7 @@ from typing import Tuple, Optional
 
 import torch
 from torch import Tensor
-if float(torch.__version__.split('.')[0]) == 0 or (float(torch.__version__.split('.')[0]) == 1 and float(torch.__version__.split('.')[1])) < 9:
+if int(torch.__version__.split('.')[0]) == 0 or (int(torch.__version__.split('.')[0]) == 1 and int(torch.__version__.split('.')[1]) < 9):
     from torch.nn.modules.linear import _LinearWithBias
 else:
     from torch.nn.modules.linear import NonDynamicallyQuantizableLinear as _LinearWithBias
